requirements.txt
Mako
psutil
python-dateutil
PyYAML
pycron
Flask
Werkzeug
Jinja2
Flask-Login
Flask-Migrate
WTForms
Flask-WTF
Flask-SQLAlchemy
SQLAlchemy
email_validator
gunicorn
flask-restx
python-dotenv
Flask-Minify

# For HTTP/HTTPS
# requests
# certifi
# For HTTPS on windows
# python-certifi-win32

# MySql and MariaDb
# mysql

# For REDIS
# redis